move packages, clean up nits
