+++ pkg/multicluster/config_handler.go
@@ -4,8 +4,7 @@ import (
 	"context"
 	"sync"
 
-	"github.com/solo-io/solo-kit/pkg/errors"
-
+	"github.com/solo-io/go-utils/errors"
 	"github.com/solo-io/go-utils/errutils"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
 	v1 "github.com/solo-io/solo-kit/pkg/multicluster/v1"
